organize: split info a few files
