+++ .gitignore
@@ -1,1 +1,2 @@
 extism/extism
+extism-dev/extism-dev
