+++ extism-dev/dev.go
@@ -1,193 +1,23 @@
 package main
 
 import (
-	_ "embed"
-	"encoding/json"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"regexp"
-	"sort"
-	"strings"
 
 	"github.com/extism/cli"
 	"github.com/spf13/cobra"
 )
 
-//go:embed repos.json
-var repos []byte
-
-type repoCategory int
-
-const (
-	Runtime = iota
-	SDK
-	PDK
-	Other
-)
-
-func (s repoCategory) String() string {
-	switch s {
-	case Runtime:
-		return "runtime"
-	case SDK:
-		return "sdk"
-	case PDK:
-		return "pdk"
-	default:
-		return "other"
-	}
-}
-
-func (s *repoCategory) Parse(cat string) {
-	if cat == "runtime" {
-		*s = Runtime
-	} else if cat == "sdk" {
-		*s = SDK
-	} else if cat == "pdk" {
-		*s = PDK
-	} else {
-		*s = Other
-	}
-}
-func (s repoCategory) MarshalJSON() ([]byte, error) {
-	return json.Marshal(s.String())
-}
-
-func (s *repoCategory) UnmarshalJSON(data []byte) (err error) {
-	var cat string
-	if err := json.Unmarshal(data, &cat); err != nil {
-		return err
-	}
-	s.Parse(cat)
-	return nil
-}
-
-type repo struct {
-	Url      string       `json:"url"`
-	Category repoCategory `json:"category"`
-}
-
-func (r repo) Split() (string, string) {
-	split := strings.Split(r.Url, "/")
-	userName := split[len(split)-2]
-	repoName := split[len(split)-1]
-	if strings.HasPrefix(r.Url, "git@") {
-		x := strings.Split(userName, ":")
-		userName = x[1]
-	}
-	return userName, repoName
-}
-
-var defaultRepos []repo
-
-func init() {
-	if err := json.Unmarshal(repos, &defaultRepos); err != nil {
-		panic(err)
-	}
-}
-
-type devArgs struct {
-	args []string
-	root string
-}
-
 type extismData struct {
 	Repos []repo `json:"repos"`
-}
-
-func (a devArgs) loadDataFile() (*extismData, error) {
-	p := filepath.Join(a.root, ".extism.dev.json")
-	cli.Log("Loading data file from", p)
-	f, err := os.Open(p)
-	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-
-	var out extismData
-	if err := json.NewDecoder(f).Decode(&out); err != nil {
-		return nil, err
-	}
-
-	return &out, nil
-}
-
-func (a devArgs) saveDataFile(data *extismData) error {
-	p := filepath.Join(a.root, ".extism.dev.json")
-	cli.Log("Saving data file to", p)
-	f, err := os.Create(p)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	enc := json.NewEncoder(f)
-	enc.SetIndent("", "  ")
-	return enc.Encode(*data)
-}
-
-func (a devArgs) link() error {
-	dest := filepath.Join(homeDir(), ".extism.dev")
-	abs, err := filepath.Abs(a.root)
-	if err != nil {
-		return err
-	}
-
-	os.Remove(dest)
-	cli.Print("Creating symlink from", abs, "to", dest)
-	return os.Symlink(abs, dest)
-}
-
-func (args *devArgs) mergeRepos(data *extismData) {
-	m := map[string]repo{}
-	for _, r := range defaultRepos {
-		m[r.Url] = r
-	}
-
-	for _, r := range data.Repos {
-		m[r.Url] = r
-	}
-
-	data.Repos = []repo{}
-	for _, v := range m {
-		data.Repos = append(data.Repos, v)
-	}
-	sort.Slice(data.Repos, func(a, b int) bool {
-		if data.Repos[a].Category != data.Repos[b].Category {
-			return data.Repos[a].Category < data.Repos[b].Category
-		}
-
-		return data.Repos[a].Url < data.Repos[b].Url
-	})
-}
-
-func (a *devArgs) SetArgs(args []string) {
-	a.args = args
 }
 
 type devInitArgs struct {
 	devArgs
 	parallel int
 	noLink   bool
-}
-
-func (repo repo) clone(root string) {
-	cli.Log("Initializing", repo.Url)
-	userName, repoName := repo.Split()
-	os.MkdirAll(userName, 0o755)
-
-	full := filepath.Join(root, userName, repoName)
-	cli.Print("Cloning", repo.Url, "to", full)
-	_, err := os.Stat(full)
-	if err == nil {
-		cli.Print("Warning:", repo.Url, "already exists")
-		return
-	}
-
-	cli.Log("Running git clone", repo.Url, full)
-	if err := exec.Command("git", "clone", repo.Url, full).Run(); err != nil {
-		cli.Print("Warning: git clone", repo.Url, "failed:", err)
-	}
 }
 
 func runDevInit(cmd *cobra.Command, args *devInitArgs) error {
@@ -250,9 +80,10 @@ func runDevEach(cmd *cobra.Command, each *devEachArgs) error {
 				continue
 			}
 		}
-		userName, repoName := repo.Split()
+		userName, repoName := repo.split()
 		p := filepath.Join(each.root, userName, repoName)
 		cli.Log("Executing", each.args[0], "in", p, "using", each.shell)
+		cli.Print(p)
 		cmd := exec.Command(each.shell, "-c", each.args[0])
 		cmd.Dir = p
 		cmd.Stdout = os.Stdout
